requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
distlib==0.3.1
Django==3.1.2
django-annoying==0.10.6
django-cors-headers==3.5.0
django-suit==2.0a1
djangorestframework==3.12.1
filelock==3.0.12
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
virtualenv==20.0.35
wrapt==1.12.1
